+++ examples/basic/variables.tf
@@ -1,0 +1,5 @@
+variable "subscription_id" {
+  type        = string
+  description = "Subscription ID where the module is deployed"
+
+}